Add math dir
