+++ intersection.h
@@ -1,7 +1,7 @@
 #ifndef _INTERSECTION_H_
 #define _INTERSECTION_H_
 
-#include "vec3.h"
+#include "math/vec3.h"
 #include "constant.h"
 
 namespace tsukihi {
