+++ math/vec4.h
@@ -4,9 +4,9 @@
 #include <cmath>
 #include <algorithm>
 
-#include "math.h"
-#include "vec2.h"
-#include "vec3.h"
+#include "math/math.h"
+#include "math/vec2.h"
+#include "math/vec3.h"
 
 namespace tsukihi {
