+++ object/point_light.h
@@ -3,7 +3,7 @@
 
 #include <cmath>
 
-#include "vec3.h"
+#include "math/vec3.h"
 #include "ray.h"
 #include "material.h"
 #include "object/object.h"
