+++ object/raymarching_mbox.h
@@ -4,8 +4,8 @@
 #include <cmath>
 #include <algorithm>
 
-#include "vec3.h"
-#include "vec4.h"
+#include "math/vec3.h"
+#include "math/vec4.h"
 #include "material.h"
 #include "constant.h"
 #include "intersection.h"
