+++ object/raymarching_menger_sponge.h
@@ -4,8 +4,8 @@
 #include <cmath>
 #include <algorithm>
 
-#include "vec3.h"
-#include "vec2.h"
+#include "math/vec3.h"
+#include "math/vec2.h"
 #include "material.h"
 #include "constant.h"
 #include "intersection.h"
