+++ object/raymarching_sphere.h
@@ -3,7 +3,7 @@
 
 #include <cmath>
 
-#include "vec3.h"
+#include "math/vec3.h"
 #include "material.h"
 #include "constant.h"
 #include "intersection.h"
